Hard code the list of past to current publisher IDs
They shouldn't change now, so this is ok to do.
And this data won't be in any API's soon, so make them static now.

Put message on home page so it's clear.
If anyone contacts to ask, either
A) it's still true they haven't changed and they can just be told it's fine
B) there may be new ones in some new system, and work can be done to surface those here or elsewhere